+++ gamemodes/terrortown/entities/weapons/weapon_tttbase.lua
@@ -241,102 +241,116 @@ local ttt2_hold_aim = CLIENT and CreateConVar("ttt2_hold_aim", 0, FCVAR_ARCHIVE,
 
 -- crosshair
 if CLIENT then
-	local GetPTranslation = LANG.GetParamTranslation
 	local TryT = LANG.TryTranslation
+	local ParT = LANG.GetParamTranslation
+
 	local mathRound = math.Round
 
 	---
 	-- @realm client
-	local sights_opacity = CreateConVar("ttt_ironsights_crosshair_opacity", "0.8", FCVAR_ARCHIVE)
+	local cvOpacitySights = CreateConVar("ttt_ironsights_crosshair_opacity", "0.8", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_size = CreateConVar("ttt_crosshair_size", "1.0", FCVAR_ARCHIVE)
+	local cvSizeCrosshair = CreateConVar("ttt_crosshair_size", "1.0", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local enable_crosshair = CreateConVar("ttt_enable_crosshair", "1", FCVAR_ARCHIVE)
+	local cvEnableCrosshair = CreateConVar("ttt_enable_crosshair", "1", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local enable_gap_crosshair = CreateConVar("ttt_crosshair_gap_enable", "0", FCVAR_ARCHIVE)
+	local cvEnableCrosshairGap = CreateConVar("ttt_crosshair_gap_enable", "0", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_gap = CreateConVar("ttt_crosshair_gap", "0", FCVAR_ARCHIVE)
+	local cvSizeCrosshairGap = CreateConVar("ttt_crosshair_gap", "0", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_opacity = CreateConVar("ttt_crosshair_opacity", "1", FCVAR_ARCHIVE)
+	local cvOpacityCrosshair = CreateConVar("ttt_crosshair_opacity", "1", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_static = CreateConVar("ttt_crosshair_static", "0", FCVAR_ARCHIVE)
+	local cvCrosshairUseWeaponscale = CreateConVar("ttt_crosshair_weaponscale", "1", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_weaponscale = CreateConVar("ttt_crosshair_weaponscale", "1", FCVAR_ARCHIVE)
+	local cvThicknessCrosshair = CreateConVar("ttt_crosshair_thickness", "1", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_thickness = CreateConVar("ttt_crosshair_thickness", "1", FCVAR_ARCHIVE)
+	local cvThicknessOutlineCrosshair = CreateConVar("ttt_crosshair_outlinethickness", "0", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local crosshair_outlinethickness = CreateConVar("ttt_crosshair_outlinethickness", "0", FCVAR_ARCHIVE)
+	local cvEnableCrosshairDot = CreateConVar("ttt_crosshair_dot", "0", FCVAR_ARCHIVE)
 
 	---
 	-- @realm client
-	local enable_dot_crosshair = CreateConVar("ttt_crosshair_dot", "0", FCVAR_ARCHIVE)
-
-	---
-	-- @realm client
-	local enable_crosshair_lines = CreateConVar("ttt_crosshair_lines", "1", FCVAR_ARCHIVE)
+	local cvEnableCrosshairLines = CreateConVar("ttt_crosshair_lines", "1", FCVAR_ARCHIVE)
 
 	local materialKeyLMB = Material("vgui/ttt/hudhelp/lmb")
 	local materialKeyRMB = Material("vgui/ttt/hudhelp/rmb")
+
+	local animData = {
+		timeStart = 0,
+		timeEnd = 0,
+		valueStart = 0,
+		valueEnd = 0
+	}
 
 	---
 	-- @see https://wiki.facepunch.com/gmod/WEAPON:DrawHUD
 	-- @realm client
 	function SWEP:DrawHUD()
-		if self.HUDHelp then
+		if isfunction(self.HUDHelp) then
 			self:DrawHelp()
 		end
 
+		if not cvEnableCrosshair:GetBool() then return end
+
 		local client = LocalPlayer()
 
-		if
-			not enable_crosshair:GetBool()
-			or not IsValid(client)
-			or SPRINT:IsSprinting(client) and not SPRINT.convars.showCrosshair:GetBool()
-		then
-			return
-		end
-
 		local sights = not self.NoSights and self:GetIronsights()
-		local x = math.floor(ScrW() * 0.5)
-		local y = math.floor(ScrH() * 0.5)
-		local scale = crosshair_weaponscale:GetBool() and math.max(0.2, 10 * self:GetPrimaryCone()) or 1
-		local timescale = 1
 
-		if not crosshair_static:GetBool() then
-			timescale = (2 - math.Clamp((CurTime() - self:LastShootTime()) * 5, 0.0, 1.0))
+		local xCenter = ScrW() * 0.5
+		local yCenter = ScrH() * 0.5
+		local scale = appearance.GetGlobalScale()
+		local scaleWeapon = cvCrosshairUseWeaponscale:GetBool() and math.max(0.2, 10 * self:GetPrimaryCone()) or 1
+		local timescale = 2 - math.Clamp((CurTime() - self:LastShootTime()) * 5, 0.0, 1.0)
+
+		-- handle size animation
+		if scaleWeapon ~= animData.valueEnd then
+			animData = {
+				timeStart = CurTime(),
+				timeEnd = CurTime() + 0.25,
+				valueStart = animData.valueEnd,
+				valueEnd = scaleWeapon
+			}
 		end
 
-		local alpha = sights and sights_opacity:GetFloat() or crosshair_opacity:GetFloat()
-		local gap = enable_gap_crosshair:GetBool() and math.floor(timescale * crosshair_gap:GetFloat()) or math.floor(20 * scale * timescale * (sights and 0.8 or 1))
-		local thickness = crosshair_thickness:GetFloat()
-		local outline = math.floor(crosshair_outlinethickness:GetFloat())
-		local length = math.floor(gap + 25 * crosshair_size:GetFloat() * scale * timescale)
+		scaleWeapon = Lerp(
+			math.ease.OutQuint(math.TimeFraction(animData.timeStart, animData.timeEnd, CurTime())),
+			animData.valueStart,
+			animData.valueEnd
+		)
+
+		local alpha = sights and cvOpacitySights:GetFloat() or cvOpacityCrosshair:GetFloat()
+		local gap = cvEnableCrosshairGap:GetBool()
+			and (timescale * cvSizeCrosshairGap:GetFloat() * scale)
+			or (20 * scaleWeapon * timescale * self:GetPrimaryConeFactor() * scale)
+		local thickness = cvThicknessCrosshair:GetFloat() * scale
+		local outline = cvThicknessOutlineCrosshair:GetFloat() * scale
+		local length = gap + 25 * cvSizeCrosshair:GetFloat() * scaleWeapon * timescale * scale
 		local offset = thickness * 0.5
 
 		if outline > 0 then
 			surface.SetDrawColor(0, 0, 0, 255 * alpha)
-			surface.DrawRect(x - length - outline, y - offset - outline, length - gap + outline * 2, thickness + outline * 2)
-			surface.DrawRect(x + gap - outline, y - offset - outline, length - gap + outline * 2, thickness + outline * 2)
-			surface.DrawRect(x - offset - outline, y - length - outline, thickness + outline * 2, length - gap + outline * 2)
-			surface.DrawRect(x - offset - outline, y + gap - outline, thickness + outline * 2, length - gap + outline * 2)
+			surface.DrawRect(xCenter - length - outline, yCenter - offset - outline, length - gap + outline * 2, thickness + outline * 2)
+			surface.DrawRect(xCenter + gap - outline, yCenter - offset - outline, length - gap + outline * 2, thickness + outline * 2)
+			surface.DrawRect(xCenter - offset - outline, yCenter - length - outline, thickness + outline * 2, length - gap + outline * 2)
+			surface.DrawRect(xCenter - offset - outline, yCenter + gap - outline, thickness + outline * 2, length - gap + outline * 2)
 		end
 
 		-- set up crosshair color
@@ -352,16 +366,16 @@ if CLIENT then
 		)
 
 		-- draw crosshair dot
-		if enable_dot_crosshair:GetBool() then
-			surface.DrawRect(x - thickness * 0.5, y - thickness * 0.5, thickness, thickness)
+		if cvEnableCrosshairDot:GetBool() then
+			surface.DrawRect(xCenter - thickness * 0.5, yCenter - thickness * 0.5, thickness, thickness)
 		end
 
 		-- draw crosshair lines
-		if enable_crosshair_lines:GetBool() then
-			surface.DrawRect(x - length, y - offset, length - gap, thickness)
-			surface.DrawRect(x + gap, y - offset, length - gap, thickness)
-			surface.DrawRect(x - offset, y - length, thickness, length - gap)
-			surface.DrawRect(x - offset, y + gap, thickness, length - gap)
+		if cvEnableCrosshairLines:GetBool() then
+			surface.DrawRect(xCenter - length, yCenter - offset, length - gap, thickness)
+			surface.DrawRect(xCenter + gap, yCenter - offset, length - gap, thickness)
+			surface.DrawRect(xCenter - offset, yCenter - length, thickness, length - gap)
+			surface.DrawRect(xCenter - offset, yCenter + gap, thickness, length - gap)
 		end
 	end
 
@@ -512,8 +526,8 @@ if CLIENT then
 		if translate then
 			extraKeyParams = extraKeyParams or {}
 			translate_params = table.Merge(extraKeyParams, defaultKeyParams)
-			primary = primary and GetPTranslation(primary, translate_params)
-			secondary = secondary and GetPTranslation(secondary, translate_params)
+			primary = primary and ParT(primary, translate_params)
+			secondary = secondary and ParT(secondary, translate_params)
 		end
 
 		--find mouse keys in the texts to add respective icons
@@ -624,14 +638,18 @@ function SWEP:PrimaryAttack(worldsnd)
 		sound.Play(self.Primary.Sound, self:GetPos(), self.Primary.SoundLevel)
 	end
 
-	self:ShootBullet(self.Primary.Damage, self.Primary.Recoil, self.Primary.NumShots, self:GetPrimaryCone())
+	self:ShootBullet(self.Primary.Damage, self:GetPrimaryRecoil(), self.Primary.NumShots, self:GetPrimaryCone())
 	self:TakePrimaryAmmo(1)
 
 	local owner = self:GetOwner()
 
 	if not IsValid(owner) or owner:IsNPC() or not owner.ViewPunch then return end
 
-	owner:ViewPunch(Angle(util.SharedRandom(self:GetClass(), -0.2, -0.1, 0) * self.Primary.Recoil, util.SharedRandom(self:GetClass(), -0.1, 0.1, 1) * self.Primary.Recoil, 0))
+	owner:ViewPunch(Angle(
+		util.SharedRandom(self:GetClass(), -0.2, -0.1, 0) * self:GetPrimaryRecoil(),
+		util.SharedRandom(self:GetClass(), -0.1, 0.1, 1) * self:GetPrimaryRecoil(),
+		0
+	))
 end
 
 ---
@@ -705,10 +723,8 @@ function SWEP:ShootBullet(dmg, recoil, numbul, cone)
 	self:GetOwner():MuzzleFlash()
 	self:GetOwner():SetAnimation(PLAYER_ATTACK1)
 
-	local sights = self:GetIronsights()
-
 	numbul = numbul or 1
-	cone = cone or 0.01
+	cone = cone or 0.02
 
 	local bullet = {}
 	bullet.Num = numbul
@@ -729,11 +745,9 @@ function SWEP:ShootBullet(dmg, recoil, numbul, cone)
 	-- Owner can die after firebullets
 	if not IsValid(self:GetOwner()) or self:GetOwner():IsNPC() or not self:GetOwner():Alive() then return end
 
-	if game.SinglePlayer() and SERVER
-	or not game.SinglePlayer() and CLIENT and IsFirstTimePredicted() then
-		-- reduce recoil if ironsighting
-		recoil = sights and (recoil * 0.6) or recoil
-
+	if SERVER and game.SinglePlayer()
+		or CLIENT and not game.SinglePlayer() and IsFirstTimePredicted()
+	then
 		local eyeang = self:GetOwner():EyeAngles()
 		eyeang.pitch = eyeang.pitch - recoil
 
@@ -744,11 +758,61 @@ end
 ---
 -- @return number
 -- @realm shared
-function SWEP:GetPrimaryCone()
-	local cone = self.Primary.Cone or 0.2
+function SWEP:GetPrimaryConeFactor()
+	local owner = self:GetOwner()
 
-	-- 10% accuracy bonus when sighting
-	return self:GetIronsights() and (cone * 0.85) or cone
+	if not IsValid(owner) then
+		return 1
+	end
+
+	if SPRINT:IsSprinting(owner) and not owner:IsOnGround() then
+		return 2.0
+	elseif SPRINT:IsSprinting(owner) or not owner:IsOnGround() then
+		return 1.6
+	elseif self:GetIronsights() then
+		return 0.8
+	else
+		return 1
+	end
+end
+
+---
+-- @return number
+-- @realm shared
+function SWEP:GetPrimaryRecoilFactor()
+	local owner = self:GetOwner()
+
+	if not IsValid(owner) then
+		return 1
+	end
+
+	if SPRINT:IsSprinting(owner) and not owner:IsOnGround() then
+		return 2.8
+	elseif SPRINT:IsSprinting(owner) or not owner:IsOnGround() then
+		return 2.2
+	elseif self:GetIronsights() then
+		return 0.6
+	else
+		return 1
+	end
+end
+
+---
+-- @return number
+-- @realm shared
+function SWEP:GetPrimaryCone()
+	local cone = self.Primary.Cone or 0.02
+
+	return cone * self:GetPrimaryConeFactor()
+end
+
+---
+-- @return number
+-- @realm shared
+function SWEP:GetPrimaryRecoil()
+	local recoil = self.Primary.Recoil or 1.5
+
+	return recoil * self:GetPrimaryRecoilFactor()
 end
 
 ---
@@ -779,6 +843,7 @@ function SWEP:SecondaryAttack()
 
 	self:SetIronsights(bNotIronsights)
 	self:SetZoom(bNotIronsights)
+
 	self:SetNextSecondaryFire(CurTime() + 0.3)
 end
 
@@ -802,6 +867,7 @@ function SWEP:Reload()
 	if self:Clip1() == self.Primary.ClipSize or self:GetOwner():GetAmmoCount(self.Primary.Ammo) <= 0 then return end
 
 	self:DefaultReload(self.ReloadAnim)
+
 	self:SetIronsights(false)
 	self:SetZoom(false)
 end
@@ -1059,8 +1125,8 @@ function SWEP:Initialize()
 	elseif SERVER then
 		self.fingerprints = {}
 
-		self:SetIronsights(false)
 		self:SetZoom(false)
+		self:SetIronsights(false)
 	end
 
 	self:SetDeploySpeed(self.DeploySpeed)