+++ gamemodes/terrortown/entities/weapons/weapon_zm_rifle.lua
@@ -91,8 +91,8 @@ end
 ---
 -- @ignore
 function SWEP:PreDrop()
-	self:SetZoom(false)
 	self:SetIronsights(false)
+	self:SetZoom(false)
 
 	return self.BaseClass.PreDrop(self)
 end
@@ -103,6 +103,7 @@ function SWEP:Reload()
 	if self:Clip1() == self.Primary.ClipSize or self:GetOwner():GetAmmoCount(self.Primary.Ammo) <= 0 then return end
 
 	self:DefaultReload(ACT_VM_RELOAD)
+
 	self:SetIronsights(false)
 	self:SetZoom(false)
 end