+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -311,6 +311,10 @@ function GM:Initialize()
 	ShopEditor.SetupShopEditorCVars()
 	ShopEditor.CreateShopDBs()
 
+	-- register synced player variables
+	player.RegisterSettingOnServer("enable_dynamic_fov", "bool")
+	player.RegisterSettingOnServer("fov_desired", "float")
+
 	-- Force friendly fire to be enabled. If it is off, we do not get lag compensation.
 	RunConsoleCommand("mp_friendlyfire", "1")
 
@@ -1360,6 +1364,10 @@ function GM:OnReloaded()
 	-- set the default random playermodel
 	self.playermodel = playermodels.GetRandomPlayerModel()
 	self.playercolor = COLOR_WHITE
+
+	-- register synced player variables
+	player.RegisterSettingOnServer("enable_dynamic_fov", "bool")
+	player.RegisterSettingOnServer("fov_desired", "float")
 
 	---
 	-- @realm shared