+++ lua/terrortown/lang/fr.lua
@@ -1046,7 +1046,6 @@ L.label_crosshair_ironsight_opacity = "Opacité du réticule du viseur"
 L.label_crosshair_size = "Taille du réticule"
 L.label_crosshair_thickness = "Épaisseur du réticule"
 L.label_crosshair_thickness_outline = "Épaisseur du contour du réticule"
-L.label_crosshair_static_enable = "Activer le réticule statique"
 L.label_crosshair_dot_enable = "Activer le point du réticule "
 L.label_crosshair_lines_enable = "Activer les lignes du réticule"
 L.label_crosshair_scale_enable = "Activer la dépendance du réticule par rapport au type d'arme"
@@ -1763,7 +1762,6 @@ L.karma_unknown_tooltip = "Inconnu"
 --L.label_sprint_max = "Max sprinting stamina"
 --L.label_sprint_stamina_consumption = "Stamina consumption factor"
 --L.label_sprint_stamina_regeneration = "Stamina regeneration factor"
---L.label_sprint_crosshair = "Show crosshair while sprinting"
 --L.label_crowbar_unlocks = "Primary attack can be used as interaction (i.e. unlocking)"
 --L.label_crowbar_pushforce = "Crowbar push force"
 
@@ -2100,6 +2098,18 @@ L.body_confirm_one = "{finder} a confirmé la mort de {victim}."
 --L.equipmenteditor_desc_builtin = "This equipment is built-in, it comes with TTT2!"
 --L.help_roles_builtin = "This role is built-in, it comes with TTT2!"
 --L.header_equipment_info = "Equipment information"
+
+-- 2023-12-22
+--L.submenu_gameplay_accessibility_title = "Accessibility"
+
+--L.header_accessibility_settings = "Accessibility Settings"
+
+--L.label_enable_dynamic_fov = "Enable dynamic FOV change"
+--L.label_enable_bobbing = "Enable view bobbing"
+--L.label_enable_bobbing_strafe = "Enable view bobbing when strafing"
+
+--L.help_enable_dynamic_fov = "Dynamic FOV is applied depending on the payer's speed. When a player is sprinting for example, the FOV is increased to visualize the speed."
+--L.help_enable_bobbing_strafe = "View bobbing is the slight camera shake while walking, swimming or falling."
 
 -- 2023-12-20
 --L.equipmenteditor_desc_damage_scaling = [[Multiplies the base damage value of a weapon by this factor.